test : Number 단위 테스트 클래스 변경
